+++ onegov/town/utils.py
@@ -155,7 +155,8 @@ class ReservationInfo(object):
             'date': self.date,
             'time': self.time,
             'delete': self.delete_link,
-            'quota': self.reservation.quota
+            'quota': self.reservation.quota,
+            'created': self.reservation.created.isoformat()
         }
 
 
@@ -313,7 +314,7 @@ class AllocationEventInfo(object):
             'wholeDay': self.allocation.whole_day,
             'partlyAvailable': self.allocation.partly_available,
             'quota': self.allocation.quota,
-            'quotaLeft': self.allocation.quota_left,
+            'quotaLeft': self.quota_left,
             'className': self.event_class,
             'partitions': self.allocation.availability_partitions(),
             'actions': [